refactor(formatter): split command helpers by responsibility
